requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
plotly==5.18.0
altair==5.2.0
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
xlsxwriter==3.1.9
requests==2.31.0
requests-cache==1.1.1
python-dateutil==2.8.2
stqdm==0.0.5
validators==0.22.0